Testcase to reproduce slow rendering
This patch adds types and code to render mails in a similar fashion Purebred
does in order to reproduce the slow performance we documented in:

https://github.com/purebred-mua/purebred/issues/208

Note, the actions are hard coded for a specific widgets. They used to be
hardcoded to perform actions on the list of threads, but with this patch they
now only perform list actions on the list of mails.